Move 'Spinner' widget class into 'widget' module.
