+++ aiidalab/app.py
@@ -20,7 +20,7 @@ from watchdog.events import FileSystemEventHandler
 from jinja2 import Template
 
 from .config import AIIDALAB_DEFAULT_GIT_BRANCH
-from .widgets import StatusHTML
+from .widgets import StatusHTML, Spinner
 from .git_util import GitManagedAppRepo as Repo
 from .utils import throttled
 
@@ -64,28 +64,6 @@ class VersionSelectorWidget(ipw.VBox):
     @traitlets.observe('disabled')
     def _observe_disabled(self, change):
         self.release_line.disabled = change['new']
-
-
-class Spinner(ipw.HTML):
-    """Widget that shows a simple spinner if enabled."""
-
-    enabled = traitlets.Bool()
-
-    def __init__(self, spinner_style=None):
-        self.spinner_style = f' style="{spinner_style}"' if spinner_style else ''
-        super().__init__()
-
-    @traitlets.default('enabled')
-    def _default_enabled(self):  # pylint: disable=no-self-use
-        return False
-
-    @traitlets.observe('enabled')
-    def _observe_enabled(self, change):
-        """Show spinner if enabled, otherwise nothing."""
-        if change['new']:
-            self.value = f"""<i class="fa fa-spinner fa-pulse"{self.spinner_style}></i>"""
-        else:
-            self.value = ""
 
 
 class AiidaLabApp(traitlets.HasTraits):
